Refactor AssetCard component by extracting vote counting and gem color logic into utility functions. Updated gem color determination to include current user ID for improved accuracy in rendering based on ownership and votes.
